Refactor tzlocal.__init__
