+++ CHANGES.txt
@@ -4,7 +4,7 @@ Changes
 1.5 (unreleased)
 ----------------
 
-- Nothing changed yet.
+- Refactor and simplify tzlocal.__init__
 
 
 1.4 (2017-04-18)
